move randomString function into module
